pyproject.toml
[project]
name = "livepodium"
version = "0.1.0"
description = "Netherlands concert discovery & ticketing platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.0",
    "dj-database-url>=3.1.0",
    "django>=6.0",
    "django-htmx>=1.27.0",
    "mollie-api-python>=3.9.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "qrcode[pil]>=8.2",
    "redis>=7.1.0",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "django-browser-reload>=1.21.0",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
